create IncomingMessage in its own module
